online_asr_with_vad_onnx/requirements.txt
loguru>=0.7.0
numpy>=1.24.0
onnxruntime>=1.22.0
onnxruntime-genai>=0.12.1
pyyaml>=6.0.0
rich>=13.7.0
sounddevice>=0.4.6
soundfile>=0.12.1
